ml_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
mlflow
sentence-transformers
supabase
pydantic
transformers
faiss-cpu